first test passed

+++ gauss.h
@@ -1,8 +1,8 @@
 #ifndef GAUSS_H
 #define GAUSS_H
 
-void triangularisation(int **first_members, int *second_member, int dim);
+void triangularisation(float **first_members, float *second_member, int dim);
 void resolution();
-void print(int **matrix, int dim);
+void print(float **matrix, int dim);
 
 #endif
